Now which all pickers are available is not configured in the coordinator_config file. Instead it is dynamically addressed. i.e. whenever a /car_events/get_states topic with a new user_id which starts with "picker" comes, that picker id is added to be tracked in the PickerStateMonitor.  That means the picker's localisation subscribers are initialised only after the picker is added to the list of pickers.
If the callarobot interface is started fresh and a new picker logs, a new /car_events/get_states topic is not published until the picker makes a call. However, as the picker monitoring is dynamic, no details of the picker is available. Earlier these cases were ignored. In this commit, the picker is added to the list of pickers and the picker's call is reset as picker is not yet localised in the PickerStateMonitor.