battery voltage check removed from RasberryCoordinator
cooridnator was checking robot battery while it is idle and when it falls below a threshold sets it as unhealthy. Similarly it was possible to set it healthy. This functionality can now be done with the unregister and register service calls. robot health monitoring and charging coorindation should be outside the task coordinator.