Removal of Specific Agents, introduction of TaskInterfaces
Agents are all managed abstractly, and all task-specific information for the time being is managed in the new InterfaceDEf classes added to custom_task modules. Now, agents can be spawned with multiple interfaces for the different roles thay may have in different tasks. There is NO major task-specific information remaining in the coordinator. If the transportation.py task definition module is removed and replaced with a working uv.py module, the coordinator will work for that task alone.